Big code move.
