+++ src/main/java/com/nexmo/client/verify/endpoints/SearchClient.java
@@ -19,14 +19,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.verify.sdk.endpoints;
+package com.nexmo.client.verify.endpoints;
 
 import com.nexmo.common.LegacyClient;
 import com.nexmo.client.NexmoResponseParseException;
 import com.nexmo.common.util.XmlUtil;
-import com.nexmo.verify.sdk.BaseResult;
-import com.nexmo.verify.sdk.SearchResult;
-import com.nexmo.verify.sdk.VerifyResult;
+import com.nexmo.client.verify.BaseResult;
+import com.nexmo.client.verify.SearchResult;
+import com.nexmo.client.verify.VerifyResult;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
